move app config provider
