+++ src/app/(pre-login)/contact/components/form.tsx
@@ -17,7 +17,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { useAppConfigContext } from '@/app/components/ConfigProvider';
+import { useAppConfigContext } from '@/app/hooks/AppProvider';
 import { zodResolver } from '@hookform/resolvers/zod';
 import {
 	Button,
