+++ src/app/components/Footer.tsx
@@ -17,12 +17,12 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import { useAppConfigContext } from '@/app/hooks/AppProvider';
 import { css, useTheme } from '@/lib/emotion';
 import { Icon, Link } from '@icgc-argo/uikit';
 import { Fragment } from 'react';
 import { Col, Row } from 'react-grid-system';
 import urljoin from 'url-join';
-import { useAppConfigContext } from './ConfigProvider';
 
 const getNavLinks = (docsUrl: string, argoRootUrl: string) => {
 	const privacyUrl = urljoin(argoRootUrl, '/page/2/privacy');
