+++ src/app/components/Hero.tsx
@@ -17,10 +17,10 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import { useAppConfigContext } from '@/app/hooks/AppProvider';
 import { BUTTON_SIZES, Button, Icon, Typography, css, useTheme } from '@icgc-argo/uikit';
 import Link from 'next/link';
 import galaxyImage from '../../../public/assets/galaxy.png';
-import { useAppConfigContext } from './ConfigProvider';
 import RegionBanner from './RegionBanner';
 
 const Hero = () => {
