+++ src/app/page.tsx
@@ -18,12 +18,12 @@
  */
 'use client';
 
+import { useAppConfigContext } from '@/app/hooks/AppProvider';
 import { css, useTheme } from '@/lib/emotion';
 import { DataCallout, Link, Typography, overtureLogo } from '@icgc-argo/uikit';
 import Image from 'next/image';
 import { ComponentType } from 'react';
 import urlJoin from 'url-join';
-import { useAppConfigContext } from './components/ConfigProvider';
 import Hero from './components/Hero';
 
 const OvertureBanner: ComponentType = () => {
